requirements.txt
Django==1.10.6
appdirs==1.4.2
djangorestframework==3.5.4
packaging==16.8
pyparsing==2.1.10
six==1.10.0
